+++ src/integrationTest/java/com/mongodb/kafka/connect/mongodb/MongoKafkaTestCase.java
@@ -22,17 +22,19 @@ import static org.junit.jupiter.api.Assertions.assertIterableEquals;
 
 import java.time.Duration;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 import java.util.Properties;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 import io.confluent.connect.avro.AvroConverter;
 
 import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
+import org.apache.kafka.common.serialization.BytesDeserializer;
+import org.apache.kafka.common.serialization.Deserializer;
 import org.apache.kafka.common.utils.Bytes;
 import org.junit.jupiter.api.extension.RegisterExtension;
 import org.slf4j.Logger;
@@ -110,12 +112,39 @@ public class MongoKafkaTestCase {
     return isMaster.get("maxWireVersion", 0) > 6;
   }
 
-  public void assertProduced(final String topicName, final int expectedCount) {
-    List<Integer> expected = IntStream.range(1, expectedCount).boxed().collect(Collectors.toList());
-    AtomicInteger counter = new AtomicInteger();
-    List<Integer> produced =
-        getProduced(topicName, b -> counter.addAndGet(1), expected, DEFAULT_MAX_RETRIES);
-    assertIterableEquals(expected, produced);
+  public MongoDatabase getDatabaseWithPostfix() {
+    return getMongoClient()
+        .getDatabase(format("%s%s", getDatabaseName(), POSTFIX.incrementAndGet()));
+  }
+
+  public MongoCollection<Document> getAndCreateCollection() {
+    MongoDatabase database = getDatabaseWithPostfix();
+    database.createCollection("coll");
+    return database.getCollection("coll");
+  }
+
+  public void assertCollection(
+      final MongoCollection<Document> source, final MongoCollection<Document> destination) {
+    assertCollection(source.find().into(new ArrayList<>()), destination);
+  }
+
+  public void assertCollection(
+      final List<Document> expected, final MongoCollection<Document> destination) {
+    int counter = 0;
+    int retryCount = 0;
+    while (retryCount < DEFAULT_MAX_RETRIES) {
+      counter++;
+      // Wait at least 3 minutes for the first data to come in.
+      if (counter > 90) {
+        retryCount++;
+      }
+      if (retryCount != DEFAULT_MAX_RETRIES && destination.countDocuments() < expected.size()) {
+        sleep(2000);
+      } else {
+        assertIterableEquals(expected, destination.find().into(new ArrayList<>()));
+        break;
+      }
+    }
   }
 
   public void assertProduced(
@@ -198,18 +227,47 @@ public class MongoKafkaTestCase {
     assertIterableEquals(docs, produced);
   }
 
+  private static final Deserializer<Bytes> BYTES_DESERIALIZER = new BytesDeserializer();
+
   public <T> List<T> getProduced(
       final String topicName,
       final Function<Bytes, T> mapper,
       final List<T> expected,
       final int maxRetryCount) {
+    return getProduced(
+        topicName,
+        BYTES_DESERIALIZER,
+        new MappingDeserializer<>(mapper),
+        ConsumerRecord::value,
+        expected,
+        maxRetryCount);
+  }
+
+  public static class MappingDeserializer<T> implements Deserializer<T> {
+    private final Function<Bytes, T> mapper;
+
+    public MappingDeserializer(final Function<Bytes, T> mapper) {
+      this.mapper = mapper;
+    }
+
+    @Override
+    public T deserialize(final String topic, final byte[] data) {
+      return mapper.apply(BYTES_DESERIALIZER.deserialize(topic, data));
+    }
+  }
+
+  public <K, V, T> List<T> getProduced(
+      final String topicName,
+      final Deserializer<K> keyDeserializer,
+      final Deserializer<V> valueDeserializer,
+      final Function<ConsumerRecord<K, V>, T> mapper,
+      final List<T> expected,
+      final int maxRetryCount) {
     LOGGER.info("Subscribing to {}", topicName);
 
-    try (KafkaConsumer<?, ?> consumer = createConsumer()) {
+    try (KafkaConsumer<K, V> consumer = createConsumer(keyDeserializer, valueDeserializer)) {
       consumer.subscribe(singletonList(topicName));
       List<T> data = new ArrayList<>();
-      T firstExpected = expected.isEmpty() ? null : expected.get(0);
-      T lastExpected = expected.isEmpty() ? null : expected.get(expected.size() - 1);
       int counter = 0;
       int retryCount = 0;
       int previousDataSize;
@@ -220,13 +278,13 @@ public class MongoKafkaTestCase {
         consumer
             .poll(Duration.ofSeconds(2))
             .records(topicName)
-            .forEach((r) -> data.add(mapper.apply((Bytes) r.value())));
+            .forEach(c -> data.add(mapper.apply(c)));
 
-        int firstExpectedIndex = data.lastIndexOf(firstExpected);
-        int lastExpectedIndex = data.lastIndexOf(lastExpected);
-        int dataSize = lastExpectedIndex - firstExpectedIndex + 1;
-        if (firstExpectedIndex > -1 && lastExpectedIndex > -1 && dataSize == expected.size()) {
-          return data.subList(firstExpectedIndex, lastExpectedIndex + 1);
+        if (data.size() >= expected.size()) {
+          int startIndex = Collections.indexOfSubList(data, expected);
+          if (startIndex > -1) {
+            return data.subList(startIndex, startIndex + expected.size());
+          }
         }
 
         // Wait at least 3 minutes for the first set of data to arrive
@@ -240,6 +298,11 @@ public class MongoKafkaTestCase {
   }
 
   public KafkaConsumer<?, ?> createConsumer() {
+    return createConsumer(null, null);
+  }
+
+  public <K, V> KafkaConsumer<K, V> createConsumer(
+      final Deserializer<K> keyDeserializer, final Deserializer<V> valueDeserializer) {
     Properties props = new Properties();
     props.put(ConsumerConfig.GROUP_ID_CONFIG, "testAssertProducedConsumer");
     props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, KAFKA.bootstrapServers());
@@ -253,7 +316,7 @@ public class MongoKafkaTestCase {
     props.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, false);
     props.put(ConsumerConfig.ISOLATION_LEVEL_CONFIG, "read_committed");
 
-    return new KafkaConsumer<>(props);
+    return new KafkaConsumer<>(props, keyDeserializer, valueDeserializer);
   }
 
   public void addSinkConnector(final String topicName) {